refact : 데이터 만드는 작업 분리하기.
