+++ book/refactoring2/ch01/createStatementData.js
@@ -1,46 +1,16 @@
-function statement(invoice, plays) {
+export default function createStatementData(invoice, plays) {
   const statementData = {};
   statementData.customer = invoice.customer;
   statementData.performances = invoice.performances.map(enrichPerformance);
   statementData.totalAmount = totalAmount(statementData);
   statementData.totalVolumeCredits = totalVolumeCredits(statementData);
-  return renderPlainText(statementData, plays);
+  return statementData;
 
   function enrichPerformance(aPerformance) {
     const result = Object.assign({}, aPerformance);
     result.play = playFor(result);
     result.amount = amountFor(result);
     result.volumeCredits = volumeCreditsFor(result);
-    return result;
-  }
-
-  function renderPlainText(data, plays) {
-    let result = `청구 내역(고객명: ${data.customer})\n`;
-
-    for (let perf of data.performances) {
-      result += `  ${perf.play.name}: ${usd(perf.amount)} (${perf.audience}석)\n`;
-    }
-
-    result += `총액: ${usd(data.totalAmount)}\n`;
-    result += `적립 포인트: ${data.totalVolumeCredits}점\n`;
-    return result;
-  }
-
-  function totalAmount(data) {
-    return data.performances
-            .reduce((total, p) => total + p.amount, 0);
-  }
-
-  function totalVolumeCredits(data) {
-    return data.performances
-            .reduce((total, p) => total + p.volumeCredits, 0);
-  }
-
-  function volumeCreditsFor(aPerformance) {
-    let result = 0;
-    result += Math.max(aPerformance.audience - 30, 0);
-    if ("comedy" == aPerformance.play.type)
-      result += Math.floor(aPerformance.audience / 5);
     return result;
   }
 
@@ -69,12 +39,22 @@ function statement(invoice, plays) {
     }
     return result;
   }
-}
 
-function usd(aNumber) {
-  return new Intl.NumberFormat("en-US",
-                    { "style": "currency", currency: "USD",
-                      minimumFractionDigits: 2 }).format(aNumber/100);
-}
+  function volumeCreditsFor(aPerformance) {
+    let result = 0;
+    result += Math.max(aPerformance.audience - 30, 0);
+    if ("comedy" == aPerformance.play.type)
+      result += Math.floor(aPerformance.audience / 5);
+    return result;
+  }
 
-module.exports = { statement };
+  function totalAmount(data) {
+    return data.performances
+            .reduce((total, p) => total + p.amount, 0);
+  }
+  
+  function totalVolumeCredits(data) {
+    return data.performances
+            .reduce((total, p) => total + p.volumeCredits, 0);
+  }
+}
