+++ book/refactoring2/ch01/index.js
@@ -1,5 +1,5 @@
-const fs = require('fs');
-const { statement } = require('./invoice'); // invoice.js에서 함수를 가져옴
+import fs from 'fs';
+import statement from './statement.js'; // invoice.js에서 함수를 가져옴
 
 // JSON 파일을 동기적으로 읽고 파싱
 const invoice = JSON.parse(fs.readFileSync('invoices.json', 'utf8'));
